Refactor Iterable module: rename AuthResponseCallback to IterableAuthResponseResult, move IterableAttributionInfo and IterableCommerceItem to separate files, and update imports accordingly
