+++ src/__mocks__/MockRNIterableAPI.ts
@@ -1,4 +1,4 @@
-import { IterableAttributionInfo } from '../Iterable';
+import { IterableAttributionInfo } from '../IterableAttributionInfo';
 import { IterableInAppMessage } from '../inApp';
 
 export class MockRNIterableAPI {
